requirements.txt
aniso8601==9.0.1
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
docopt==0.6.2
Flask==2.2.2
flask-restx==0.5.1
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-dialogflow==2.15.1
google-cloud-dialogflow-cx==1.13.1
google-cloud-speech==2.15.1
googleapis-common-protos==1.56.4
grpcio==1.48.1
grpcio-status==1.48.1
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
pipreqs==0.4.11
playsound==1.3.0
proto-plus==1.22.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pydub==0.25.1
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.2.1
requests==2.28.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
yarg==0.1.9
